Add AdminLTE style
